+++ test/chain/vm_test.dart
@@ -468,8 +468,11 @@ void main() {
       });
 
       expect(chain.traces, hasLength(2));
+
+      // Assert that we've trimmed the VM's stack chains here to avoid
+      // duplication.
       expect(chain.traces.first.toString(),
-          equals(new Trace.from(trace).toString()));
+          equals(new Chain.parse(trace.toString()).traces.first.toString()));
       expect(
           chain.traces.last.frames, contains(frameMember(startsWith('main'))));
     });